websocket: Add live market data stream via Yahoo WebSockets.
This commit marks the major addition of live market data for real time
market research.

We use a Web worker to connect to the Yahoo websocket endpoint and
subscribe to the relevant tickers. This allows for a dedicated thread
for live data processing, thus reducing any performance bottlenecks in
the main thread. The websocket then emits protobuf-encoded live stream
which is then decoded and parsed into JSON via the protobuf.js library.

The pricing data obtained through the standard API call to the Yahoo
finance public API is used to initialize the live state object in React
which is then replaced with the live market data from websocket stream.

Note: The "proto/yfinance.proto" is compiled into "utils/yfinance.js"
and "utils/yfinance.d.ts" via the protobufjs-cli tool.